chore(quality): CI и чистый прогон ruff/mypy (F1/F2)
CI (.github/workflows/ci.yml): backend — ruff, mypy, pytest без маркера integration; frontend — lint, typecheck, test, build. Инструменты добавлены в requirements-dev.txt, конфиг — backend/pyproject.toml (набор правил ruff с подписанными исключениями, mypy с check_untyped_defs; 43 файла, 0 ошибок).

Не косметика, а находки: ASYNC230 в api/uploads.py (запись EDF ушла в asyncio.to_thread), SQLAlchemy 2.0 (DeclarativeBase, async_sessionmaker) в models/db.py, JobState с валидацией статуса, восстановленного из файла задачи, cast MGHImage в nibabel-вызовах, raise ... from exc в роутах, удаление дублей Field(env=...) в core/config.py.

AGENTS.md и docs/rules/tests.md — команды статических проверок и раздел о них.

+++ backend/tests/conftest.py
@@ -99,7 +99,7 @@ def write_minimal_edf(path, ch_names, data_uv, sfreq, record_sec=1.0):
     """
     ns = len(ch_names)
     n_times = data_uv.shape[1]
-    samps_per_record = int(round(sfreq * record_sec))
+    samps_per_record = round(sfreq * record_sec)
     n_records = int(np.ceil(n_times / samps_per_record))
     pad = n_records * samps_per_record - n_times
     if pad:
